file - renamed folder for supabase defaults to defaults, autoformatting
