+++ app/auth/callback/route.ts
@@ -1,17 +1,17 @@
-import { createClient } from '@/utils/supabase/default/server';
-import { NextResponse } from 'next/server';
-import { cookies } from 'next/headers';
+import {createClient} from '@/utils/supabase/defaults/server';
+import {NextResponse} from 'next/server';
+import {cookies} from 'next/headers';
 
 export async function GET(request: Request) {
-  const requestUrl = new URL(request.url);
-  const code = requestUrl.searchParams.get('code');
+	const requestUrl = new URL(request.url);
+	const code = requestUrl.searchParams.get('code');
 
-  if (code) {
-    const cookieStore = cookies();
-    const supabase = createClient(cookieStore);
-    await supabase.auth.exchangeCodeForSession(code);
-  }
+	if (code) {
+		const cookieStore = cookies();
+		const supabase = createClient(cookieStore);
+		await supabase.auth.exchangeCodeForSession(code);
+	}
 
-  // URL to redirect to after sign in process completes
-  return NextResponse.redirect(requestUrl.origin);
+	// URL to redirect to after sign in process completes
+	return NextResponse.redirect(requestUrl.origin);
 }
